backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
drf-spectacular>=0.27,<1.0
psycopg2-binary>=2.9,<3.0
celery>=5.3,<6.0
redis>=5.0,<6.0
# weasyprint>=60.0,<70.0  # Requires GTK - use only with Docker
Pillow>=10.0,<11.0
python-dotenv>=1.0,<2.0
gunicorn>=21.0,<23.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0
pytest-xdist>=3.5,<4.0

# Code quality
flake8>=6.1,<7.0
black>=23.0,<24.0
isort>=5.12,<6.0
pylint-django>=2.5,<3.0